+++ doc/source/ReleaseNotes.rst
@@ -1,18 +1,19 @@
 
 Release History
 ===============
-This contains the release notes for the current version of nptsne.
+This contains the release notes for the current version (|version|) of nptsne.
 
 .. _release-notes-label:
 
 Release notes
 -------------
 
+.. include:: ./changelogs/v1.0.0rc1.rst
 .. include:: ./changelogs/v0.2.0.rst
 .. include:: ./changelogs/v0.1.1a1.rst
 
 Previous versions
 -----------------
 
+v0.2.0
 v0.1.1a1
-